requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
boto3==1.35.84
pydantic==2.10.3
asyncio-throttle==1.0.2
python-multipart==0.0.20
jinja2==3.1.4
aiofiles==24.1.0
pandas==2.2.3
numpy==2.2.1
xlsxwriter==3.2.0
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
requests==2.31.0
python-dotenv==1.0.0
slowapi==0.1.9
